Changed migration while to use float instead of double

+++ db/migrate/20130315040025_add_latitude_longitude_columns_to_events.rb
@@ -1,6 +1,6 @@
 class AddLatitudeLongitudeColumnsToEvents < ActiveRecord::Migration
   def change
-    add_column :events, :latitude, :double
-    add_column :events, :longitude, :double
+    add_column :events, :latitude, :float
+    add_column :events, :longitude, :float
   end
 end
